Assistant checkpoint: Fix all errors + finalize medium-term + begin long-term
Assistant generated file changes:
- tests/lums.test.js: Fix Jest ESM module configuration
- server/lums/vorax.c: Fix multi-character constant warnings
- server/lums/operations.c: Fix signed/unsigned comparison warnings, Fix LUMStructureType comparison warning
- package.json: Add missing test script
- Makefile: Enhanced Makefile with strict warnings
- server/vm-vir.ts: Implement V-IR Virtual Machine
- server/linear-types.ts: Implement Linear Type System
- server/routes.ts: Integrate linear type checking, Add linear type validation to operations
- server/vorax-bootstrap.ts: Begin long-term bootstrap implementation
- RAPPORT_FINAL_LUMS_VORAX_V6.md: Create final comprehensive report V6

---

User prompt:

Je détecter encore des erreurs, de fois les mêmes ?!Explication de ces erreurs et corrige et Vérifie que tu a bien réalisé ce que je t'es demandé : tu est un expert senior critique dans les domaines nécessaires que tu dois m’indiquer les quel

Cette règle est aussi spécialement pour toi même ! Exigences de lecture complète

"à chaque vérification, modification / correction / mise à jour du code source, l'ensemble du code (dossier → sous-dossier → fichier → sous-fichier → ligne) devra être relu/inspecté" et applique immédiatement tes recommandations à cours terme qui ce trouve dans le RAPPORT_ANALYSE_LUMS_VORAX.md et V2 et V3, v4 v5ainsi que tout les règles décrites et présente moi ton tout nouveau rapport final dan un nouveau md sans toucher au ancien âpre test, vérification et validation sans erreur 3 fois et quel phase somme nous réellement ? Cours terme , moyen terme du RAPPORT_ANALYSE_LUMS_VORAX.md et V2 et V3, v4, v5 ? Je veux la vérité avec preuves test réel d'authenticité. Continue et finalise la phase moyen terme et début la long terme si possible déjà immédiatement.

+++ server/lums/operations.c
@@ -54,7 +54,7 @@ LUMGroup** lum_split(LUMGroup* source, int zones, size_t* result_count) {
     size_t current_index = 0;
     
     for (int i = 0; i < zones; i++) {
-        size_t zone_count = lums_per_zone + (i < remainder ? 1 : 0);
+        size_t zone_count = lums_per_zone + ((size_t)i < remainder ? 1 : 0);
         
         if (zone_count == 0) {
             // Create empty zone
@@ -276,7 +276,7 @@ LUMGroup** lum_contextual_split(LUMGroup* source, size_t* result_count) {
         
         size_t type_index = 0;
         for (size_t i = 0; i < source->count; i++) {
-            if (source->lums[i].structure_type == type) {
+            if ((int)source->lums[i].structure_type == (int)type) {
                 type_lums[type_index] = source->lums[i];
                 type_lums[type_index].position.x = type_index * 20;
                 type_lums[type_index].position.y = type * 40;
